Move bench/ under script/
