--- bench/output/macrobase.bench.PrecisionRecallPipelineTest/telematicsSimple/06-10-01:39:48/results0.txt
@@ -1,6 +1,0 @@
-Welcome to
-  _   _   _   _   _   _   _   _   _  
- / \ / \ / \ / \ / \ / \ / \ / \ / \ 
-( m | a | c | r | o | b | a | s | e )
- \_/ \_/ \_/ \_/ \_/ \_/ \_/ \_/ \_/ 
-
