--- bench/output/macrobase.bench.SampleComparePipeline/campaignExpendituresComplex/06-07-10:17:37/batch.conf
@@ -1,34 +1,0 @@
-macrobase.query.name: campaignExpendituresComplex
-macrobase.pipeline.class: macrobase.analysis.pipeline.BasicBatchedPipeline
-
-macrobase.loader.attributes: [cand_nm, contbr_nm, contbr_zip, contbr_employer, contbr_employer]
-macrobase.loader.targetLowMetrics: []
-macrobase.loader.targetHighMetrics: [contb_receipt_amt]
-
-macrobase.analysis.transformType: MAD_OR_MCD
-
-macrobase.loader.db.user: pbailis
-macrobase.loader.db.password: None
-macrobase.loader.db.baseQuery: SELECT * FROM campaign_expenditures LIMIT 100000;
-
-macrobase.analysis.minSupport: 0.0010
-macrobase.analysis.minOIRatio: 3.0000
-
-macrobase.analysis.useZScore: false
-macrobase.analysis.zscore.threshold: 3.0000
-
-macrobase.analysis.usePercentile: true
-macrobase.analysis.targetPercentile: 0.9900
-
-macrobase.analysis.mcd.alpha: 0.5000
-macrobase.analysis.mcd.stoppingDelta: 0.0010
-
-macrobase.loader.loaderType: CACHING_POSTGRES_LOADER
-macrobase.loader.db.cacheDirectory: sql_cache
-
-logging:
-  level: INFO
-
-  loggers:
-    "macrobase": DEBUG
-macrobase.pipeline.class: macrobase.bench.SampleComparePipeline
