--- bench/output/macrobase.bench.SampleComparePipeline/campaignExpendituresComplex/06-07-10:17:37/results0.txt
@@ -1,23 +1,0 @@
-Welcome to
-  _   _   _   _   _   _   _   _   _  
- / \ / \ / \ / \ / \ / \ / \ / \ / \ 
-( m | a | c | r | o | b | a | s | e )
- \_/ \_/ \_/ \_/ \_/ \_/ \_/ \_/ \_/ 
-
-Exception in thread "main" java.lang.ClassNotFoundException: macrobase.bench.SampleComparePipeline
-	at java.net.URLClassLoader$1.run(URLClassLoader.java:372)
-	at java.net.URLClassLoader$1.run(URLClassLoader.java:361)
-	at java.security.AccessController.doPrivileged(Native Method)
-	at java.net.URLClassLoader.findClass(URLClassLoader.java:360)
-	at java.lang.ClassLoader.loadClass(ClassLoader.java:424)
-	at sun.misc.Launcher$AppClassLoader.loadClass(Launcher.java:308)
-	at java.lang.ClassLoader.loadClass(ClassLoader.java:357)
-	at java.lang.Class.forName0(Native Method)
-	at java.lang.Class.forName(Class.java:259)
-	at macrobase.runtime.command.MacroBasePipelineCommand.run(MacroBasePipelineCommand.java:27)
-	at macrobase.runtime.command.MacroBasePipelineCommand.run(MacroBasePipelineCommand.java:15)
-	at io.dropwizard.cli.ConfiguredCommand.run(ConfiguredCommand.java:77)
-	at io.dropwizard.cli.Cli.run(Cli.java:70)
-	at io.dropwizard.Application.run(Application.java:80)
-	at macrobase.runtime.MacroBaseServer.main(MacroBaseServer.java:21)
-	at macrobase.MacroBase.main(MacroBase.java:33)
