--- bench/output/macrobase.bench.SampleComparePipeline/campaignExpendituresComplex/06-07-10:21:18/results0.txt
@@ -1,45 +1,0 @@
-Welcome to
-  _   _   _   _   _   _   _   _   _  
- / \ / \ / \ / \ / \ / \ / \ / \ / \ 
-( m | a | c | r | o | b | a | s | e )
- \_/ \_/ \_/ \_/ \_/ \_/ \_/ \_/ \_/ 
-
-DEBUG [2016-06-07 17:21:20,802] macrobase.analysis.pipeline.BasePipeline: Running query campaignExpendituresComplex
-DEBUG [2016-06-07 17:21:20,808] macrobase.analysis.pipeline.BasePipeline: CONFIG:
-macrobase.analysis.mcd.alpha: 0.5
-macrobase.analysis.mcd.stoppingDelta: 0.001
-macrobase.analysis.minOIRatio: 3.0
-macrobase.analysis.minSupport: 0.001
-macrobase.analysis.targetPercentile: 0.99
-macrobase.analysis.transformType: MAD_OR_MCD
-macrobase.analysis.usePercentile: true
-macrobase.analysis.useZScore: false
-macrobase.analysis.zscore.threshold: 3.0
-macrobase.loader.attributes: cand_nm,contbr_nm,contbr_zip,contbr_employer,contbr_employer
-macrobase.loader.db.baseQuery: SELECT * FROM campaign_expenditures LIMIT 100000;
-macrobase.loader.db.cacheDirectory: sql_cache
-macrobase.loader.db.password: None
-macrobase.loader.db.user: pbailis
-macrobase.loader.loaderType: CACHING_POSTGRES_LOADER
-macrobase.loader.targetHighMetrics: contb_receipt_amt
-macrobase.loader.targetLowMetrics: 
-macrobase.pipeline.class: macrobase.bench.SampleComparePipeline
-macrobase.query.name: campaignExpendituresComplex
-INFO  [2016-06-07 17:21:20,821] macrobase.ingest.DiskCachingIngester: Data did not exist; going to read from SQL.
-INFO  [2016-06-07 17:21:21,601] macrobase.ingest.DiskCachingIngester: Writing out loaded data...
-INFO  [2016-06-07 17:21:22,458] macrobase.ingest.DiskCachingIngester: ...done writing!
-INFO  [2016-06-07 17:21:22,943] macrobase.conf.MacroBaseConf: By default: using MAD transform for dimension 1 metric.
-INFO  [2016-06-07 17:21:23,101] macrobase.conf.MacroBaseConf: By default: using MAD transform for dimension 1 metric.
-DEBUG [2016-06-07 17:21:23,145] macrobase.bench.SampleComparePipeline: Sample size is 10
-DEBUG [2016-06-07 17:21:23,346] macrobase.analysis.classify.BatchingPercentileClassifier: 0.99 Percentile Cutoff: 136.1111111111122
-DEBUG [2016-06-07 17:21:23,349] macrobase.analysis.classify.BatchingPercentileClassifier: Median: 2.2222222222222503
-DEBUG [2016-06-07 17:21:23,350] macrobase.analysis.classify.BatchingPercentileClassifier: Max: 3375.000000000027
-Exception in thread "main" java.lang.NullPointerException
-	at macrobase.bench.SampleComparePipeline.run(SampleComparePipeline.java:121)
-	at macrobase.runtime.command.MacroBasePipelineCommand.run(MacroBasePipelineCommand.java:35)
-	at macrobase.runtime.command.MacroBasePipelineCommand.run(MacroBasePipelineCommand.java:15)
-	at io.dropwizard.cli.ConfiguredCommand.run(ConfiguredCommand.java:77)
-	at io.dropwizard.cli.Cli.run(Cli.java:70)
-	at io.dropwizard.Application.run(Application.java:80)
-	at macrobase.runtime.MacroBaseServer.main(MacroBaseServer.java:20)
-	at macrobase.MacroBase.main(MacroBase.java:33)
