+++ script/bench/batch_template.conf
