+++ script/bench/conf/campaigns-1M.yaml
