+++ script/bench/conf/dashboard_config.json
