+++ script/bench/conf/istc_config.json
