+++ script/bench/conf/plotting_config.json
