+++ script/bench/experiments/3gaussians-ok.yaml
