+++ script/bench/produce_plots.py
