+++ script/bench/remove_exception_msgs_from_output_log.py
