+++ script/bench/streaming_template.conf
