+++ script/bench/synthetic/1d_mcd_killer-kde.yaml
