+++ script/bench/synthetic/2d_mcd_killer.yaml
